+++ src/main/java/com/besson/arknights/block/custom/fantastic/BrownBeanBagSofa.java
@@ -27,7 +27,7 @@ public class BrownBeanBagSofa extends FurnitureHorizontalFacingBlock {
     @Override
     public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
         if (!world.isClient){
-            return SeatEntity.create(world,pos,0.25,player,state.get(FACING));
+            return SeatEntity.create(world,pos,0.05,player,state.get(FACING));
         }
         return ActionResult.SUCCESS;
     }